% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CarreraLuz22}
\alias{cafe01}
\title{Tiempo de carrera la luz Cali 2022}
\format{
Data con 1922 registros y 7 variables
\describe{
\item{id}{Identificador único del cporredor.}
\item{sex}{sexo del corredor.}
\item{edad}{edad del correodor.}
\item{origen}{ciudad de origen de lcorredor.}
\item{timerun}{tiempo de carrera en segundos.}
\item{categoria}{categoria del corredor.}
\item{nacionalidad}{nacionalidad del corredor.}
}
}
\usage{
  cafe01
}
\description{
  Esta data contiene información sobre los tiempo y principales 
  caracteristicas de los participantes en la carrera la 
  Luz 2022 en Cali - Valle del Cauca Colombia .
}
\examples{
  data(CarreraLuz22)
}
\keyword{datasets}
